// solar-dashboard-frontend/src/app/+locations/models/location.ts
/**
 * This is Dealer interface to give data type for Dealer object
 */

export interface ILocation {
  
 locNbr: string;
 contactPerson:string;
  locAddrName: string;
  address: string;
  city: string;
 state: string;
 country: string;
  zipCode: string;
  email: string;
  phoneNumber: string;
  latitude: string;
   longitude: string;
  createdTS: string;
  lastUpdatedTS: string;
  
 
}